work on docs

+++ savu/plugins/reconstructions/tomobar/tomobar_recon_3D_tools.py
@@ -22,7 +22,7 @@ toolbox and the CCPi RGL toolkit.
 
         data_Huber_thresh:
             visibility: advanced
-            dtype: [None,int]
+            dtype: [None,float]
             description: "Threshold parameter for __Huber__ data fidelity."
             default: None
